+++ src/token.cpp
@@ -26,23 +26,15 @@ auto operator<<(std::ostream& os, Token const& t) -> std::ostream& {
 }
 
 auto get_type_from_lexeme(std::string const& str) -> std::optional<TokenType> {
-    auto const lookup_map = std::map<std::string, TokenType>{{"fn", TokenType::FN},
-                                                             {"using", TokenType::USING},
-                                                             {"as", TokenType::AS},
-                                                             {"i64", TokenType::TYPE},
-                                                             {"i8", TokenType::TYPE},
-                                                             {"if", TokenType::IF},
-                                                             {"else", TokenType::ELSE},
-                                                             {"else if", TokenType::ELSE_IF},
-                                                             {"bool", TokenType::TYPE},
-                                                             {"void", TokenType::TYPE},
-                                                             {"mut", TokenType::MUT},
-                                                             {"let", TokenType::LET},
-                                                             {"return", TokenType::RETURN},
-                                                             {"extern", TokenType::EXTERN},
-                                                             {"while", TokenType::WHILE},
-                                                             {"true", TokenType::TRUE},
-                                                             {"false", TokenType::FALSE}};
+    auto const lookup_map = std::map<std::string, TokenType>{
+        {"fn", TokenType::FN},         {"using", TokenType::USING},     {"as", TokenType::AS},
+        {"i64", TokenType::TYPE},      {"i32", TokenType::TYPE},        {"i8", TokenType::TYPE},
+        {"u64", TokenType::TYPE},      {"u32", TokenType::TYPE},        {"u8", TokenType::TYPE},
+        {"f64", TokenType::TYPE},      {"f32", TokenType::TYPE},        {"if", TokenType::IF},
+        {"else", TokenType::ELSE},     {"else if", TokenType::ELSE_IF}, {"bool", TokenType::TYPE},
+        {"void", TokenType::TYPE},     {"mut", TokenType::MUT},         {"let", TokenType::LET},
+        {"return", TokenType::RETURN}, {"extern", TokenType::EXTERN},   {"while", TokenType::WHILE},
+        {"true", TokenType::TRUE},     {"false", TokenType::FALSE}};
 
     return lookup_map.find(str) != lookup_map.end() ? std::make_optional(lookup_map.at(str)) : std::nullopt;
 }
@@ -90,6 +82,7 @@ auto operator<<(std::ostream& os, TokenType const& t) -> std::ostream& {
     case TokenType::ELSE_IF: os << "ELSE_IF"; break;
     case TokenType::AMPERSAND: os << "AMPERSAND"; break;
     case TokenType::UNSIGNED_INTEGER: os << "UNSIGNED_INTEGER"; break;
+    case TokenType::FLOAT_LITERAL: os << "FLOAT_LITERAL"; break;
     default: os << "UNKNOWN";
     }
     return os;